Extracted check MOJO
To make goal `spotless:check` accessible. Moved all functionality
common for "apply" and "check" into `AbstractSpotlessMojo`. Now command
`mvn spotless:check` will result in all formatting violations being
printed and build being failed.
