+++ plugin-maven/src/main/java/com/diffplug/maven/spotless/SpotlessApplyMojo.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2016 DiffPlug
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.diffplug.maven.spotless;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
+
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugins.annotations.Mojo;
+
+import com.diffplug.spotless.Formatter;
+
+@Mojo(name = "apply")
+public class SpotlessApplyMojo extends AbstractSpotlessMojo {
+
+	@Override
+	protected void process(List<File> files, Formatter formatter) throws MojoExecutionException {
+		for (File file : files) {
+			try {
+				formatter.applyTo(file);
+			} catch (IOException e) {
+				throw new MojoExecutionException("Unable to format file " + file, e);
+			}
+		}
+	}
+}
